Refactor with deleting matplotlib
