+++ app/prepare_data.py
@@ -4,8 +4,6 @@ import numpy as np
 import pandas as pd
 from scipy.interpolate import interp1d
 
-from app.saving_plot import TransmittancePlotter
-
 class PrepareData:
 
     def __init__(self):
@@ -26,35 +24,6 @@ class PrepareData:
             answer = None
         return answer
 
-
-    def launch(self, general_data, layer_form_data, add_data, wit):
-        if wit == 'angle':  # angle dependecy
-
-            n, k, d = self.layer_angle(layer_form_data)
-            parameters = self.specific_data_angle(add_data)  # wv and angle
-
-        elif wit == 'wv':  # wavelenght dependecy
-            d, material = self.layer_wv(layer_form_data)
-            dts = [self.refractive_index_for_material(val) for val in material.values()]
-            wv, n, k = self.interpolate(dts)
-
-            angle_list = add_data['angle'].split(',')
-            parameters = {'angle': [float(angle) * (2 * np.pi) / 360 for angle in angle_list], 'wv': wv,
-                         'material': material}  # wv and angle
-
-        aol = len(d) + 1  # number of layers
-        d[aol] = [0]
-        k[0], k[aol] = 0, 0
-
-        parameters2, n_gen = self.take_general(general_data, aol)  # y_label and polarization
-        n.update(n_gen)  # n0 and n_last
-
-        parameters.update(parameters2)
-        parameters.update({'n': n, 'k': k, 'd': d, 'amountoflayers': aol, 'wit': wit})
-        parameters['title'] = self.titles(**parameters)  # from parametrs dict make title for plot
-        plotter = TransmittancePlotter(**parameters)
-        wv, output, path = plotter.run()
-        return wv, output, path
 
 
     def layer_angle(self, layer_form_data):  # Get data for angle dependency.
