+++ app/saving_plot.py
@@ -1,10 +1,7 @@
-import numpy as np
-import matplotlib.pyplot as plt
-
-import matplotlib
-matplotlib.use('agg')
-
 import time
+
+import numpy as np
+import plotly.graph_objects as go
 
 from app.layer_transmittance_calculator import LayerTransmittanceCalculator
 
@@ -20,12 +17,17 @@ class TransmittancePlotter:
     def __init__(self, **parameters):
         self.layer_params = LayerParameters(parameters)
         self.output_url = {}
-        self.ylab = {'R':'Коэффициент отражения', 'T':'Коэффициент пропускания', 'OD':'Оптическая плотность'}
+        self.fig = go.Figure()
+        self.ylab = {
+            'R':'Reflection coefficient',
+            'T':'Transmission coefficient',
+            'OD':'Optical density'}
 
 
     # height - dict for calculation; d – dict with all values of height layers, o – what layer
     # het(d = {1:[40,20,0], 2:[30,0], 3:[50,100]}, height = {0:0},  layer = 1)
-    def het(self, d, ri, height={}, o=1):
+    def het(self, d,ri, height={}, o=1):
+
         for h in d[o]:
             height[o] = float(h)
             if (h == 0) & (o != self.layer_params.amountoflayers):
@@ -33,22 +35,22 @@ class TransmittancePlotter:
                 ri['n'].pop(o)
                 ri['k'].pop(o)
             try:
-                wv, output = self.het(d, ri, height, o + 1)
+                x, output = self.het(d,ri, height, o + 1)
             except:
                 self.layer_params.height = height
                 if self.layer_params.wit == 'angle':
-                    wv, output = self.calc_on_teta(height, ri)
+                    x, output = self.calc_on_teta(height, ri)
                 else:
-                    wv, output = self.calc_on_wv(height, ri)
-        return wv, output
+                    x, output = self.calc_on_wv(height, ri)
+        return x, output
 
     def calc_on_wv(self, height, ri):
         output, wv = self.plot_transmittance_vs_wavelength(height, ri)
-        return wv, self.output_url
+        return wv, output
 
     def calc_on_teta(self, height, ri):
-        output, angle_inDegree = self.plot_transmittance_vs_angle(height, ri)
-        return angle_inDegree, self.output_url
+        angle_inDegree, output = self.plot_transmittance_vs_angle(height, ri)
+        return angle_inDegree, output
 
     def plot_transmittance_vs_wavelength(self, height, ri):
         angle = self.layer_params.angle
@@ -59,23 +61,31 @@ class TransmittancePlotter:
             output_new, wv = layer_transmittance.calculate_output()
             output[theta] = output_new
             self.label = self.labels_good()
-            plt.plot(wv, output[theta][self.layer_params.y_label], label=self.label)
-            plt.xlabel('Длина Волны, нм')
+            self.fig.add_trace(go.Scatter(
+                x=wv, y=output[theta][self.layer_params.y_label],
+                name=self.label, line_shape='linear'
+            ))
             self.output_urls(wv, output[theta])
+        self.fig.update_layout(xaxis_title='Wavelength, nm')
         return output, wv
 
     def plot_transmittance_vs_angle(self, height, ri):
         angle = self.layer_params.angle
         self.layer_params.theta = angle
+
         layer_transmittance = LayerTransmittanceCalculator(self.layer_params, height, ri)
         output, wv = layer_transmittance.calculate_output()
 
         angle_inDegree = angle * 360 / (2 * np.pi)
+
         self.label = self.labels_good()
-        plt.plot(angle_inDegree, output[self.layer_params.y_label], label=self.label)
-        plt.xlabel('Угол, градусы')
+        self.fig.add_trace(go.Scatter(
+            x=angle_inDegree, y=output[self.layer_params.y_label],
+            name=self.label,line_shape='linear'
+        ))
+        self.fig.update_layout(xaxis_title='Angle, degree')
         self.output_urls(angle_inDegree, output)
-        return output, angle_inDegree
+        return angle_inDegree, output
 
     def output_urls(self, X, output):
         save_file = np.vstack(
@@ -110,7 +120,7 @@ class TransmittancePlotter:
         return str_title
 
     def labels_good(self):  # angle, height, n0, n_last
-        labels = {}  # str??
+        labels = {}
         aol = self.layer_params.amountoflayers
 
         for key, d in self.layer_params.title['d'].items():
@@ -127,24 +137,20 @@ class TransmittancePlotter:
         return labels
 
     def run(self):
-
         aol = self.layer_params.amountoflayers
         n_last = self.layer_params.n[aol]
         d = self.layer_params.d
+
         for n_aol in n_last:
             self.layer_params.n[aol] = n_aol
             ri = {
                 'n': self.layer_params.n,
                 'k': self.layer_params.k
             }
-            wv, output = self.het(d, ri)
+            self.het(d, ri)
 
-        plt.ylabel(self.ylab[self.layer_params.y_label])
-        str_title = self.title_to_str()
-        plt.title(str_title)
-        plt.legend()  # title = parametrs['title']
-        filename_plot = str(int(time.time())) + '.png'
-        image_path = 'app/static/plot/' + filename_plot
-        plt.savefig(image_path)
-        plt.clf()  # delete plot
-        return wv, output, filename_plot
+        self.fig.update_layout(
+            title=self.title_to_str(),
+            yaxis_title=self.ylab[self.layer_params.y_label]
+        )
+        return self.fig, self.output_url
