--- templates/answer.html
@@ -1,17 +1,0 @@
-{% extends "header.html" %}
-{% block content %}
-
-	<img src="{{IMAGE}}"><p>
-	<p>
-		Скачать в формате CSV:<p>
-	{% for name,url in output_csv.items() %}	
-	<b>{{name}}:</b> <a href={{url_for('static', filename='plot/'+url)}} download>Скачать </a>
-	<p>
-        <p></p>
-	{% endfor %}
-
-<!--		<b> {{q}} </b>
-	<b> {{mat}} </b> -->
-		
-
-	{% endblock %}
