+++ templates/wv_index.html
@@ -31,8 +31,7 @@
 
 {% endfor %}
 
-  <button type="submit" class="button" formaction="{{ url_for('plotting') }}">Send (old version)</button>
-  <button type="submit" class="button" formaction="{{ url_for('plot_new') }}">Send 2(New)</button>
+  <button type="submit" class="button" formaction="{{ url_for('plotting') }}">Send</button>
 
 </form>
